Adds manual mode for rho_peri so that whatever the user puts in model is used for conductivity
Adds none mode for SuffShift so that the cuff is not moved at all (user would need to know exactly what they were doing)

+++ src/core/simulation.py
@@ -402,8 +402,11 @@ class Simulation(Exceptionable, Configurable, Saveable):
         # neuron files
         du.copy_tree(os.path.join(os.environ[Env.PROJECT_PATH.value], 'src', 'neuron'), target)
         # enums (used by submit.py in neuron files)
-        os.mkdir(os.path.join(target, 'utils'))
-        du.copy_tree(os.path.join(os.environ[Env.PROJECT_PATH.value], 'src', 'utils'), os.path.join(target, 'utils'))
+
+        utils_path = os.path.join(target, 'utils')
+        if not os.path.exists(utils_path):
+            os.mkdir(os.path.join(target, 'utils'))
+            du.copy_tree(os.path.join(os.environ[Env.PROJECT_PATH.value], 'src', 'utils'), os.path.join(target, 'utils'))
 
     @staticmethod
     def import_n_sims(sample: int, model: int, sim: int, sim_dir: str, source: str):
